requirements.txt
torch>=2.0.0
torchvision>=0.15.0
datasets>=2.14.0
huggingface-hub>=0.16.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
pillow>=10.0.0
geopandas>=0.13.0
folium>=0.14.0
plotly>=5.15.0
streamlit>=1.25.0
